requirements.txt
aws-cdk-lib
aws-cdk.aws-lambda-python-alpha
pydantic
